requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# X (Twitter) API
tweepy==4.14.0

# Environment and configuration
python-dotenv==1.0.1
pydantic-settings==2.2.1
openai>=0.28.0

# HTTP client
httpx==0.27.0
aiohttp==3.9.3

# Data processing
pandas==2.2.0

# Caching
cachetools==5.3.2

# CORS
starlette==0.36.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# YouTube API
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0

# Windows COM interface (for CeVIO AI)
pywin32>=306; sys_platform == "win32"

